requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
async-timeout==3.0.1
bidict==0.23.1
blinker==1.8.2
chardet==3.0.4
contourpy==1.3.0
cycler==0.12.1
dicom-validator==0.6.3
dnspython==2.2.1
dominate==2.9.1
easywebdav==1.2.0
eventlet==0.37.0
filelock==3.16.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SocketIO==5.4.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==0.14.2
flywheel-sdk==19.3.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gevent==24.10.3
gevent-websocket==0.10.1
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.0
imageio==2.35.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
labkey==3.2.0
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
matplotlib==3.9.2
mpmath==1.3.0
multidict==4.7.6
mysql-connector-python==9.0.0
networkx==3.3
nibabel==5.3.2
nilearn==0.10.4
nptyping==2.5.0
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pathvalidate==3.2.1
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.2
pydicom==3.0.1
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pynrrd==1.0.0
pyparsing==3.1.4
python-engineio==4.10.1
python-gdcm==3.0.24.1
python-socketio==5.11.4
redis==5.2.0
requests-toolbelt==1.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.0.0
SimpleITK==2.4.0
SQLAlchemy==2.0.35
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
torchviz==0.0.2
tzdata==2024.2
visitor==0.1.3
vtk==9.3.1
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.17.1
zope.event==5.0
zope.interface==7.1.1
